DAPI-1250: Improve retrieval of products to evaluate
